#12 Move a matcher test to the base class.
